# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kalapi Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: getAllUsers
      tags: [admin]
      summary: Get all users (admin)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/wishlist:
    get:
      operationId: getWishlist
      tags: [user]
      summary: Get user wishlist
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: toggleWishlist
      tags: [user]
      summary: Toggle wishlist item
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistRequest"
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
  /user/addresses:
    get:
      operationId: getAddresses
      tags: [user]
      summary: Get user addresses
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: addAddress
      tags: [user]
      summary: Add address
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressRequest"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /user/addresses/{id}:
    delete:
      operationId: deleteAddress
      tags: [user]
      summary: Delete address
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        image:
          type: string
        isVeg:
          type: boolean
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - category
        - isVeg
        - isAvailable
...
    Address:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        street:
          type: string
        city:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
      required:
        - id
        - street
        - city
        - pincode
    AddressRequest:
      type: object
      properties:
        label:
          type: string
        street:
          type: string
        city:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
      required:
        - street
        - city
        - pincode
    WishlistRequest:
      type: object
      properties:
        menuItemId:
          type: integer
      required:
        - menuItemId
    WishlistResponse:
      type: object
      properties:
        added:
          type: boolean
        menuItemId:
          type: integer
      required:
        - added
        - menuItemId
...